backend/requirements.txt
flask==2.3.2
flask-sqlalchemy==3.0.3
flask-marshmallow==0.15.0
marshmallow-sqlalchemy==0.28.1
flask-jwt-extended==4.4.4
python-dotenv==1.0.0
flask-cors==3.0.10
pytest==7.3.1
gunicorn==20.1.0
